09.04.2024_7

+++ app/Http/Controllers/Admin/UserController.php
@@ -66,8 +66,12 @@ class UserController extends Controller
 
     public function destroy($id)
     {
-        $user = User::find($id);
-        $user->delete();
-        return redirect(route('admin.users.index'));
+        if (User::count() > 1) {
+            $user = User::find($id);
+            $user->delete();
+            return redirect(route('admin.users.index'));
+        } else {
+            return 'Нельзя удалить единственного пользователя';
+        }
     }
 }
